requirements.txt
streamlit>=1.32.0
langchain>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
chromadb>=0.4.24
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
sentence-transformers>=2.6.0
pypdf>=4.0.0
google-generativeai>=0.8.0
numpy>=1.24.0
pytest>=8.0.0
